Update op03n10b.mscx
I do agree with the cadence at the end of the first phrase, but shouldn't we also add the the PAC sign (as well as on 5.1) instead of "waiting" till 9.1 for the first PAC? Check the changes please I've entered please.

14-15 now harmony added.

22 - same music same labeling, removed therefore one label (diminished seventh) to match 9.

Do you really hear a cadence in the middle of 31? I hear it at best as an evaded cadence swallowed by what follows it.
pretty much the same about the HC in 34 (only arriving at 38)